task 1&2 ended
